first draft of multi-target codebase
