+++ examples-native/babel.config.js
@@ -1,0 +1,6 @@
+module.exports = function(api) {
+  api.cache(true)
+  return {
+    presets: ['babel-preset-expo'],
+  }
+}
