+++ src/targets/web/index.ts
@@ -1,0 +1,2 @@
+export * from '../../shared/index'
+export * from './canvas'
