+++ src/three-types.ts
@@ -1,5 +1,7 @@
 import * as THREE from 'three'
-import { PointerEvent } from './canvas'
+// import { PointerEvent } from './canvas'
+
+type PointerEvent = null
 
 export type NonFunctionKeys<T> = { [K in keyof T]: T[K] extends Function ? never : K }[keyof T]
 export type Overwrite<T, O> = Omit<T, NonFunctionKeys<O>> & O
@@ -12,7 +14,7 @@ type Args<T> = T extends (new (...args: any) => any) ? ConstructorParameters<T> 
 export declare namespace ReactThreeFiber {
   type Euler = THREE.Euler | number[]
   type Matrix4 = THREE.Matrix4 | number[]
-  type Vector3 = THREE.Vector3 | number[]
+  type Vector3 = THREE.Vector3 | [number]
   type Color = THREE.Color | number | string
   export type Events = {
     onClick?: (e: PointerEvent) => void
