+++ tsconfig.json
@@ -14,6 +14,6 @@
     "outDir": "types",
     "resolveJsonModule": true
   },
-  "include": ["./index.ts", "/src/**/*"],
+  "include": ["./src"],
   "exclude": ["./node_modules/**/*"]
 }
